feat: create reusable container component
